fix: fullcalendar util함수 분리
